clean up and doc

+++ install-virtuoso.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
-wget https://github.com/openlink/virtuoso-opensource/archive/v7.1.0.zip -O virtuoso-opensource-7.1.0.zip
+if [ ! -f virtuoso-opensource-7.1.0.zip ]; then
+    wget https://github.com/openlink/virtuoso-opensource/archive/v7.1.0.zip -O virtuoso-opensource-7.1.0.zip
+fi
 unzip virtuoso-opensource-7.1.0.zip
 mkdir virtuoso
 cd virtuoso-opensource-7.1.0
